// src/generated/client/models/NotificationType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Lipsanen Project Management API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Notification type
 * @export
 */
export const NotificationType = {
    TaskAssigned: 'TASK_ASSIGNED',
    TaskStatusChanged: 'TASK_STATUS_CHANGED',
    ChangeProposalCreated: 'CHANGE_PROPOSAL_CREATED',
    ChangeProposalStatusChanged: 'CHANGE_PROPOSAL_STATUS_CHANGED',
    CommentLeft: 'COMMENT_LEFT'
} as const;
export type NotificationType = typeof NotificationType[keyof typeof NotificationType];


export function NotificationTypeFromJSON(json: any): NotificationType {
    return NotificationTypeFromJSONTyped(json, false);
}

export function NotificationTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): NotificationType {
    return json as NotificationType;
}

export function NotificationTypeToJSON(value?: NotificationType | null): any {
    return value as any;
}